refactor database helper and configure test scripts to run on test env
