--- src/test-setup/db-handler.js
@@ -1,36 +1,0 @@
-import mongoose from 'mongoose';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { MongoMemoryServer } from 'mongodb-memory-server';
-
-const mongod = new MongoMemoryServer();
-
-const opts = {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  reconnectTries: Number.MAX_VALUE,
-  reconnectInterval: 1000,
-};
-
-// eslint-disable-next-line import/prefer-default-export
-export const connectTestDb = async () => {
-  const mongoUri = await mongod.getUri();
-  await mongoose.connect(mongoUri, opts, err => {
-    if (err) {
-      console.error(err);
-    }
-  });
-};
-
-export const closeTestDb = async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
-  await mongod.stop();
-};
-
-export const clearTestDb = async () => {
-  const { collections } = mongoose.connection;
-  for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany();
-  }
-};
